--- plugin/docusaurus-plugin-api-extractor-markdown-documenter/src/mark/builders/code-span.ts
@@ -1,7 +1,0 @@
-import { DocCodeSpan, TSDocConfiguration } from '@microsoft/tsdoc';
-import { CodeSpanBuilder } from './interfaces';
-
-export const initCodeSpan =
-  (configuration: TSDocConfiguration): CodeSpanBuilder =>
-  (code: string) =>
-    new DocCodeSpan({ configuration, code });
