--- plugin/docusaurus-plugin-api-extractor-markdown-documenter/src/mark/builders/fenced-code.ts
@@ -1,7 +1,0 @@
-import { DocFencedCode, TSDocConfiguration } from '@microsoft/tsdoc';
-import { FencedCodeBuilder } from './interfaces';
-
-export const initFencedCode =
-  (configuration: TSDocConfiguration): FencedCodeBuilder =>
-  (code: string, language: string = 'typescript') =>
-    new DocFencedCode({ configuration, code, language });
