--- plugin/docusaurus-plugin-api-extractor-markdown-documenter/src/mark/builders/paragraph.ts
@@ -1,7 +1,0 @@
-import { DocNode, DocParagraph, TSDocConfiguration } from '@microsoft/tsdoc';
-import { ParagraphBuilder } from './interfaces';
-
-export const initParagraph =
-  (configuration: TSDocConfiguration): ParagraphBuilder =>
-  (children: DocNode[]): DocParagraph =>
-    new DocParagraph({ configuration }, children);
