--- plugin/docusaurus-plugin-api-extractor-markdown-documenter/src/mark/builders/table-row.ts
@@ -1,8 +1,0 @@
-import { TSDocConfiguration } from '@microsoft/tsdoc';
-import { DocTableCell } from '../nodes/doc-table-cell';
-import { DocTableRow } from '../nodes/doc-table-row';
-import { TableRowBuilder } from './interfaces';
-
-export const initTableRow = (configuration: TSDocConfiguration): TableRowBuilder => {
-  return (children: DocTableCell[]) => new DocTableRow({ configuration }, children);
-};
