--- plugin/docusaurus-plugin-api-extractor-markdown-documenter/src/mark/builders/text.ts
@@ -1,10 +1,0 @@
-import { DocPlainText, TSDocConfiguration } from '@microsoft/tsdoc';
-import { TextBuilder } from './interfaces';
-
-export const initText =
-  (configuration: TSDocConfiguration): TextBuilder =>
-  (text: string) =>
-    new DocPlainText({
-      configuration,
-      text
-    });
